fix: rename AdditionalRoomState file
